Fix errors when loading phrase table
Change ints to longs
Add missing constructors because of null pointer exceptions
Add test phrase table file and load it in main

+++ joshua/Vocabulary.java
@@ -2,7 +2,7 @@ package joshua;
 
 public interface Vocabulary {
 
-  int find(String word);
-  int findOrInsert(String word);
+  long find(String word);
+  long findOrInsert(String word);
 
 }
