+++ joshua/phrase/search/Note.java
@@ -2,5 +2,5 @@ package joshua.phrase.search;
 
 public class Note {
 	// *vp, *cvp, ints
-	public int[] cvp;
+	public long[] cvp;
 }
